Reorganized: Each class should have its own file.
